refactor(app): inline single-consumer component Props interfaces
Apply the documented rule: a React component's props type is named `Props`
and declared inline. Convert the 23 single-consumer `*PropsInterface`
files in /interface to inline `interface Props` in their components and
delete the interface files.

Kept the 4 genuinely shared props interfaces (consumed by 2+ components or
used as a base): transaction-actions-menu-props, rule-pill-slot-props,
transaction-filter-page-header-props, update-transaction-form-props.
