--- packages/app/src/rule/interface/matching-rules-pill-props.interface.ts
@@ -1,4 +1,0 @@
-export interface MatchingRulesPillPropsInterface {
-    readonly matchingRulesCount: number;
-    readonly matchingRuleIds: readonly number[];
-}
