--- packages/app/src/rule/interface/rules-list-page-props.interface.ts
@@ -1,4 +1,0 @@
-export interface RulesListPagePropsInterface {
-    readonly matchingRuleIds?: readonly number[];
-    readonly onGoBack: () => void;
-}
