--- packages/app/src/sync/interface/account-bank-sync-card-props.interface.ts
@@ -1,3 +1,0 @@
-export interface AccountBankSyncCardPropsInterface {
-    readonly accountId: number;
-}
