--- packages/app/src/sync/interface/resync-bank-sync-account-props.interface.ts
@@ -1,3 +1,0 @@
-export interface ResyncBankSyncAccountPropsInterface {
-    readonly accountId: number;
-}
