+++ packages/app/src/transaction/components/transaction-card-account-info/transaction-card-account-info.tsx
@@ -2,12 +2,17 @@ import { View } from 'react-native';
 
 import { isDefined } from '@rnw-community/shared';
 
+
 import { TransactionAccountLine } from '../transaction-account-line/transaction-account-line';
 import { TransactionCardSelector } from '../transaction-card/transaction-card.selector';
 
-import type { TransactionCardAccountInfoPropsInterface } from '../../interface/transaction-card-account-info-props.interface';
+import type { TransactionWithRelationsEntityInterface } from '@budgie/contracts';
 
-export const TransactionCardAccountInfo = ({ transaction }: TransactionCardAccountInfoPropsInterface) => {
+interface Props {
+    readonly transaction: TransactionWithRelationsEntityInterface;
+}
+
+export const TransactionCardAccountInfo = ({ transaction }: Props) => {
     const { toAccount, fromAccount } = transaction;
 
     if (isDefined(fromAccount) && isDefined(toAccount)) {
