--- packages/app/src/transaction/interface/consolidation-source-modal-content-props.interface.ts
@@ -1,7 +1,0 @@
-import type { EmptyFn } from '@rnw-community/shared';
-
-export interface ConsolidationSourceModalContentPropsInterface {
-    readonly transactionId: number;
-    readonly onClose: EmptyFn;
-    readonly onRevertSuccess: EmptyFn;
-}
