--- packages/app/src/transaction/interface/consolidation-source-row-props.interface.ts
@@ -1,8 +1,0 @@
-import type { ConsolidationSourceRowInterface, TransactionConsolidationTypeEnum } from '@budgie/contracts';
-
-export interface ConsolidationSourceRowPropsInterface {
-    readonly source: ConsolidationSourceRowInterface;
-    readonly index: number;
-    readonly consolidationType: TransactionConsolidationTypeEnum | null;
-    readonly testID?: string;
-}
