--- packages/app/src/transaction/interface/convert-to-refund-content-props.interface.ts
@@ -1,6 +1,0 @@
-import type { ConvertToRefundModalResolveType } from './convert-to-refund-modal-resolve.type';
-
-export interface ConvertToRefundContentPropsInterface {
-    readonly refundIncomeTransactionId: number;
-    readonly resolveConvertToRefund: ConvertToRefundModalResolveType;
-}
