--- packages/app/src/transaction/interface/convert-to-refund-footer-props.interface.ts
@@ -1,9 +1,0 @@
-import type { TransactionPickerItemInterface } from './transaction-picker-item.interface';
-
-export interface ConvertToRefundFooterPropsInterface {
-    readonly selectedCandidate: TransactionPickerItemInterface | null;
-    readonly showRevert: boolean;
-    readonly onClose: () => void;
-    readonly onConvert: () => void;
-    readonly onRevert: () => void;
-}
