--- packages/app/src/transaction/interface/refunded-pill-props.interface.ts
@@ -1,7 +1,0 @@
-import type { TransactionWithRelationsEntityInterface } from '@budgie/contracts';
-
-export interface RefundedPillPropsInterface {
-    readonly transaction: TransactionWithRelationsEntityInterface;
-    readonly onPress?: () => void;
-    readonly testID?: string;
-}
