--- packages/app/src/transaction/interface/transaction-account-line-props.interface.ts
@@ -1,8 +1,0 @@
-import type { UserIconNameEnum } from '@budgie/contracts';
-
-export interface TransactionAccountLinePropsInterface {
-    readonly direction: 'from' | 'to';
-    readonly icon: UserIconNameEnum;
-    readonly title: string;
-    readonly testID?: string;
-}
