--- packages/app/src/transaction/interface/transaction-card-account-info-props.interface.ts
@@ -1,5 +1,0 @@
-import type { TransactionWithRelationsEntityInterface } from '@budgie/contracts';
-
-export interface TransactionCardAccountInfoPropsInterface {
-    readonly transaction: TransactionWithRelationsEntityInterface;
-}
