--- packages/app/src/transaction/interface/transaction-convert-menu-item-props.interface.ts
@@ -1,8 +1,0 @@
-import type { UserIconNameEnum } from '@budgie/contracts';
-
-export interface TransactionConvertMenuItemPropsInterface {
-    readonly icon: UserIconNameEnum;
-    readonly label: string;
-    readonly onConvert: () => void;
-    readonly testID: string;
-}
