--- packages/app/src/transaction/interface/transaction-list-context-menu-props.interface.ts
@@ -1,11 +1,0 @@
-import type { PopoverMenuAnchor } from '../../@generic/component/popover-menu/popover-menu';
-import type { TransactionWithRelationsEntityInterface } from '@budgie/contracts';
-import type { EmptyFn } from '@rnw-community/shared';
-
-export interface TransactionListContextMenuPropsInterface {
-    readonly transaction: TransactionWithRelationsEntityInterface | null;
-    readonly anchor?: PopoverMenuAnchor;
-    readonly isOpen: boolean;
-    readonly onClose: EmptyFn;
-    readonly onCloseComplete: EmptyFn;
-}
