--- packages/app/src/transaction/interface/transaction-list-convert-menu-item-props.interface.ts
@@ -1,7 +1,0 @@
-import type { EmptyFn } from '@rnw-community/shared';
-
-export interface TransactionListConvertMenuItemPropsInterface {
-    readonly isVisible: boolean;
-    readonly isRefund?: boolean;
-    readonly onConvert: EmptyFn;
-}
