--- packages/app/src/transaction/interface/transaction-picker-row-props.interface.ts
@@ -1,8 +1,0 @@
-import type { TransactionPickerItemInterface } from './transaction-picker-item.interface';
-
-export interface TransactionPickerRowPropsInterface {
-    readonly item: TransactionPickerItemInterface;
-    readonly isSelected?: boolean;
-    readonly onPress?: (item: TransactionPickerItemInterface) => void;
-    readonly testID?: string;
-}
