--- packages/app/src/transaction/interface/transaction-picker-search-dock-props.interface.ts
@@ -1,9 +1,0 @@
-import type { ReactNode } from 'react';
-
-export interface TransactionPickerSearchDockPropsInterface {
-    readonly search: string;
-    readonly searchPlaceholder: string;
-    readonly footer: ReactNode;
-    readonly onSearchChange: (value: string) => void;
-    readonly searchTestID: string;
-}
