--- packages/app/src/transaction/interface/uncategorized-transactions-pill-props.interface.ts
@@ -1,4 +1,0 @@
-export interface UncategorizedTransactionsPillPropsInterface {
-    readonly count: number;
-    readonly onPress: () => void;
-}
