--- packages/app/src/transaction/interface/update-transaction-actions-menu-props.interface.ts
@@ -1,10 +1,0 @@
-import type { TransactionActionsMenuPropsInterface } from './transaction-actions-menu-props.interface';
-
-export interface UpdateTransactionActionsMenuPropsInterface extends Pick<
-    TransactionActionsMenuPropsInterface,
-    'onDelete' | 'isConsolidated'
-> {
-    readonly onRevert: () => void;
-    readonly onConvertToRefund?: () => void;
-    readonly onConvertToTransfer: () => void;
-}
